feat: refactor candidate card component into subcomponents for better readability and maintainability
